# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartCity OS — Enterprise Smart City Operations & Digital Governance Platform API
  version: 1.0.0
  description: |
    Unified event-driven platform unifying citizen portals, municipal department
    operations, spatial GIS analytics, real-time IoT sensor telemetry and automated
    emergency command centres. All requests (except auth) are authenticated with a
    Bearer JWT issued by the Auth service.
  contact:
    name: SmartCity OS
servers:
  - url: http://localhost:4000/api/v1
    description: Local API Gateway
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a citizen account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password]
              properties:
                fullName: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                phoneNumber: { type: string }
      responses:
        '201':
          description: Created — returns AuthSession (access + refresh tokens)
        '422': { description: Validation failed }
        '409': { description: Account already exists }
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain tokens
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                rememberMe: { type: boolean, default: false }
      responses:
        '200': { description: AuthSession returned }
        '401': { description: Invalid credentials }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token for fresh access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: New token pair }
  /auth/logout:
    post:
      tags: [Auth]
      security: [bearerAuth: []]
      summary: Invalidate a session
      responses:
        '200': { description: Logged out }
...
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password with a reset token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: Password reset }
...
  /auth/me:
    get:
      tags: [Auth]
      security: [bearerAuth: []]
      summary: Current authenticated user profile
      responses:
        '200': { description: AuthUser }
  /auth/sessions:
    get:
      tags: [Auth]
      security: [bearerAuth: []]
      summary: List active sessions
      responses:
        '200': { description: Array of sessions }

  # ----------------- Complaints -----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []